services/requirements.txt
connexion[swagger-ui]==2.9.0
Flask==1.1.4
Flask-Cors==3.0.10
gunicorn==20.1.0
psycopg2-binary==2.9.1
python-dotenv==0.19.2
prometheus-client==0.11.0
pytest==6.2.5
pandas==1.3.5
openpyxl==3.0.9
passlib==1.7.4
xlsxwriter
MarkupSafe==2.0.1
numpy==1.21.5
Flask-Session==0.4.0
SQLAlchemy==1.4.7
pandasql==0.7.3
plotly==5.9.0
Shapely==1.8.2  